refactor(schemas,core): host the cimd client id predicate in schemas
