+++ packages/experience/src/utils/sign-in-experience.ts
@@ -3,8 +3,7 @@
  * Remove this once we have a better way to get the sign in experience through SSR
  */
 
-import { SignInIdentifier, VerificationType } from '@logto/schemas';
-import { isCimdClientId } from '@logto/shared/universal';
+import { isCimdClientId, SignInIdentifier, VerificationType } from '@logto/schemas';
 import { isObject } from '@silverhand/essentials';
 import i18next from 'i18next';
