+++ packages/schemas/src/utils/cimd.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 
-import { isCimdClientId } from './cimd-client-id.js';
+import { isCimdClientId } from './cimd.js';
 
 describe('isCimdClientId', () => {
   it('matches the fork scheme test case-insensitively and never matches registered ids', () => {
