+++ packages/shared/src/utils/index.ts
@@ -1,4 +1,3 @@
-export * from './cimd-client-id.js';
 export * from './fetch.js';
 export * from './id.js';
 export * from './normalize-error.js';
